Prevent caching the guide in development
