+++ src/routes/guide/[lang]/index.json.js
@@ -1,19 +1,15 @@
-import data from './_data.js';
-
-const lookup = {};
-data.forEach(({ lang, sections }) => {
-	lookup[lang] = JSON.stringify(sections);
-});
+import getLanguages from './_data.js';
 
 export function get(req, res, next) {
 	const { lang } = req.params;
+	const languages = getLanguages();
 
-	if (lang in lookup) {
+	if (lang in languages) {
 		res.writeHead(200, {
 			'Content-Type': 'application/json',
 			'Cache-Control': `max-age=${30 * 60 * 1e3}` // 30 minutes
 		});
-		res.end(lookup[lang]);
+		res.end(languages[lang]);
 	} else {
 		next();
 	}
